// src/app/models/navio.ts
export class Navio {
  id: number;
  nome: string;
  tonelagem: string;
  classeDp: number;
  propulsao: string;
  comprimento: number;
  boca: number;
  calado: number;
  areaLivreConves: number;
  bollardPull: string;
  imo: number;
  callSign: string;
  anoConstrucao: Date;
  proximaDocagem: Date;
}